set feature api at features file from home page
